Add files via upload
The .svg file is the fixed high-resolution version of mapping published articles and projects pointing to various technological solutions involving science communication. The original visualization occurred via Gephi (final.gephi file).
This map is part of the report prepared for and funded by the Scicom Observatory research group of Berlin and Brandenburg Academy of Sciences.